+++ packages/message-lint-rules/messageWithoutSource/src/messageWithoutSource.ts
@@ -2,16 +2,14 @@ import type { MessageLintRule } from "@inlang/message-lint-rule"
 import { id, displayName, description } from "../marketplace-manifest.json"
 
 export const messageWithoutSourceRule: MessageLintRule = {
-	meta: {
-		id: id as MessageLintRule["meta"]["id"],
-		displayName,
-		description,
-	},
-	message: ({ message: { id, variants }, sourceLanguageTag, report }) => {
-		if (!variants.some((variant) => variant.languageTag === sourceLanguageTag)) {
+	id: id as MessageLintRule["id"],
+	displayName,
+	description,
+	run: ({ message, settings, report }) => {
+		if (!message.variants.some((variant) => variant.languageTag === settings.sourceLanguageTag)) {
 			report({
 				messageId: id,
-				languageTag: sourceLanguageTag,
+				languageTag: settings.sourceLanguageTag,
 				body: {
 					en: `Message with id '${id}' is specified, but missing in the source.`,
 				},