+++ packages/versioned-interfaces/message-lint-rule/src/index.ts
@@ -1,11 +1,1 @@
 export * from "./interface.js"
-
-/**
- * -------- RE-EXPORTS --------
- *
- * See https://github.com/inlang/monorepo/issues/1184.
- */
-
-export * from "@inlang/message"
-export * from "@inlang/language-tag"
-export * from "@inlang/translatable"