+++ packages/versioned-interfaces/marketplace-manifest/CHANGELOG.md
@@ -9,7 +9,7 @@
 ### Patch Changes
 
 - Updated dependencies [973858c6]
-  - @inlang/project-config@1.2.0
+  - @inlang/project-settings@1.2.0
   - @inlang/translatable@1.1.0
 
 ## 1.1.0
@@ -21,4 +21,4 @@
 ### Patch Changes
 
 - Updated dependencies [713385d3]
-  - @inlang/project-config@1.1.0
+  - @inlang/project-settings@1.1.0